--- packages/adapter-firebase-admin/firebase.json
@@ -1,10 +1,0 @@
-{
-  "firestore": {
-    "rules": "firestore.rules"
-  },
-  "emulator": {
-    "firestore": {
-      "port": 8080
-    }
-  }
-}
