--- packages/adapter-firebase-admin/jest.config.js
@@ -1,1 +1,0 @@
-module.exports = require("@next-auth/adapter-test/jest/jest-preset")
