--- packages/adapter-firebase-admin/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@next-auth/firebase-admin-adapter",
-  "version": "1.0.0",
-  "description": "Firebase Admin adapter for next-auth.",
-  "homepage": "https://authjs.dev",
-  "repository": "https://github.com/nextauthjs/next-auth",
-  "bugs": {
-    "url": "https://github.com/nextauthjs/next-auth/issues"
-  },
-  "author": "Wyatt Ades <me@wyattades.com>",
-  "contributors": [],
-  "main": "dist/index.js",
-  "files": [
-    "dist",
-    "index.d.ts"
-  ],
-  "license": "ISC",
-  "keywords": [
-    "next-auth",
-    "next.js",
-    "firebase",
-    "firebase-admin"
-  ],
-  "private": false,
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "tsc",
-    "test": "firebase emulators:exec --only firestore --project next-auth-admin-test jest"
-  },
-  "peerDependencies": {
-    "firebase-admin": "^11",
-    "next-auth": "^4"
-  },
-  "devDependencies": {
-    "@next-auth/adapter-test": "workspace:*",
-    "@next-auth/tsconfig": "workspace:*",
-    "firebase-admin": "^11.4.1",
-    "firebase-tools": "^11.16.1",
-    "jest": "^27.4.3",
-    "next-auth": "workspace:*"
-  }
-}
