--- packages/adapter-firebase/.firebaserc
@@ -1,1 +1,0 @@
-{}
