+++ packages/adapter-firebase/firestore.rules
