+++ packages/adapter-firebase/package.json
@@ -21,7 +21,8 @@
   "keywords": [
     "next-auth",
     "next.js",
-    "firebase"
+    "firebase",
+    "firebase-admin"
   ],
   "private": false,
   "publishConfig": {
@@ -29,16 +30,16 @@
   },
   "scripts": {
     "build": "tsc",
-    "test": "FIRESTORE_EMULATOR_HOST=localhost:8080 firebase --token '$FIREBASE_TOKEN' emulators:exec --only firestore --project next-auth-test jest"
+    "test": "firebase emulators:exec --only firestore --project next-auth-test jest"
   },
   "peerDependencies": {
-    "firebase": "^9.7.0",
+    "firebase-admin": "^11.4.1",
     "next-auth": "^4"
   },
   "devDependencies": {
     "@next-auth/adapter-test": "workspace:*",
     "@next-auth/tsconfig": "workspace:*",
-    "firebase": "^9.14.0",
+    "firebase-admin": "^11.4.1",
     "firebase-tools": "^11.16.1",
     "jest": "^27.4.3",
     "next-auth": "workspace:*"
